add wallet name support in client config
