+++ apps/coordinator/src/components/Wallet/WalletGenerator.jsx
@@ -264,7 +264,11 @@ class WalletGenerator extends React.Component {
     try {
       const { getBlockchainClient } = this.props;
       const client = await getBlockchainClient();
-      await client.getFeeEstimate();
+      if (client.bitcoindParams.walletName) {
+        await client.getWalletInfo();
+      } else {
+        await client.getFeeEstimate();
+      }
       setPasswordError("");
       this.setState({ connectSuccess: true }, () => {
         // if testConnection was passed a callback
@@ -278,7 +282,7 @@ class WalletGenerator extends React.Component {
         setPasswordError(
           "Unauthorized: Incorrect username and password combination",
         );
-      else setPasswordError(e.message);
+      else setPasswordError(e.response?.data?.error?.message || e.message);
     }
   };
